Shields Up clean up
New lines
Removed Pompano's effects file, as it was also located inside of the SU
folder
Removed the commodities file as it's outdated
Removed the hai file as it was located inside of That Someone
Removed the interface file because I have no idea what it's there for
and it was giving me errors
Removed some destinations in valhalla.txt. The comment on that one
mission that refers to assist mission only; if a mission is offered on
landing then the destination is the source planet.
Removed the Milky Way system.