I made some needed adjustments to the map file in the 000-Shields Up folder.
I moved the New New Providence and the Teshana systems to account for the latest vanilla systems added in that area. I also created a link from Teshana to Orbona, one of the new vanilla systems added in that area. Additionally, I made the planet, Dead Man's Refuge, able to be dominated; the fleet type, fleet amount, combat rating requirement, and the tribute price could always be adjusted as needed.